rename file of constants
